chore: encapsulate IPC
